+++ docs/changelog.md
@@ -16,6 +16,7 @@ This will probably be the last update for a while until I figure out raster and 
 - Added `layer.x`, `layer.y` convenience properties.
 - Tweaked `layer.position` to be relative to the top left, not bottom left.
 - Added a binding class to `layer.styles` such that `.append` (etc) work.
+- Auto-set `fillType` based on if a style has a `gradient` provided.
 - The beginnings of a `layer.adjusts` property, a library of tweaks and dials that allows, say, `layer.adjusts.white_balance.tint = -0.4`. (Not completed!)
 
 ### 0.0.4 (2020-01-15)